# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Secure Ephemeral Data Exchange Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DetailedHealthStatus {
  status: string;
  containers_created: number;
  containers_delivered: number;
  memory_used_mb: number;
  memory_free_mb: number;
  timestamp: string;
}

export interface FileMetadata {
  name: string;
  size: number;
  type: string;
  originalIndex: number;
}

export type CreateShareRequestShareType =
  (typeof CreateShareRequestShareType)[keyof typeof CreateShareRequestShareType];

export const CreateShareRequestShareType = {
  text: "text",
  files: "files",
  mixed: "mixed",
} as const;

export interface CreateShareRequest {
  /** Base64-encoded encrypted payload (ChaCha20-Poly1305) */
  encryptedData: string;
  /**
   * Time-to-live in seconds (60 to 345600)
   * @minimum 60
   * @maximum 345600
   */
  ttl: number;
  /** Optional PBKDF2-derived hash of the password (for verification) */
  passwordHash?: string | null;
  /** Base64-encoded salt used for password hashing */
  passwordSalt?: string | null;
  /** Optional HTTPS webhook URL */
  webhookUrl?: string | null;
  /** Optional custom webhook message */
  webhookMessage?: string | null;
  fileMetadata?: FileMetadata[] | null;
  shareType: CreateShareRequestShareType;
  /** Total size of the payload in bytes */
  totalSize: number;
  /** hCaptcha token (optional in dev mode) */
  captchaToken?: string | null;
}

export interface CreateShareResponse {
  shareId: string;
  /** ISO timestamp when share expires */
  expiresAt: string;
}

export type GetShareResponseShareType =
  (typeof GetShareResponseShareType)[keyof typeof GetShareResponseShareType];

export const GetShareResponseShareType = {
  text: "text",
  files: "files",
  mixed: "mixed",
} as const;

export interface GetShareResponse {
  encryptedData: string;
  fileMetadata?: FileMetadata[] | null;
  passwordRequired: boolean;
  passwordSalt?: string | null;
  shareType: GetShareResponseShareType;
  totalSize: number;
  webhookUrl?: string | null;
  webhookMessage?: string | null;
}

export type PeekShareResponseShareType =
  (typeof PeekShareResponseShareType)[keyof typeof PeekShareResponseShareType];

export const PeekShareResponseShareType = {
  text: "text",
  files: "files",
  mixed: "mixed",
} as const;

export interface PeekShareResponse {
  totalSize: number;
  passwordRequired: boolean;
  shareType: PeekShareResponseShareType;
  fileCount: number;
  expiresAt: string;
}

export interface DeleteShareResponse {
  success: boolean;
  message: string;
}

export interface TestWebhookRequest {
  webhookUrl: string;
}

export interface TestWebhookResponse {
  success: boolean;
  statusCode?: number | null;
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface RateLimitErrorResponse {
  error: string;
  message: string;
  retryAfterSeconds: number;
}

export interface ShareErrorResponse {
  error: string;
  message: string;
  humorousMessage: string;
}