# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Secure Ephemeral Data Exchange Platform API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns detailed telemetry — requires Authorization Bearer header
 * @summary Detailed health check (requires API key)
 */
export const HealthDetailedResponse = zod.object({
  status: zod.string(),
  containers_created: zod.number(),
  containers_delivered: zod.number(),
  memory_used_mb: zod.number(),
  memory_free_mb: zod.number(),
  timestamp: zod.string(),
});

/**
 * Upload encrypted data and create a one-time share link
 * @summary Create a new share
 */
export const createShareBodyTtlMin = 60;
export const createShareBodyTtlMax = 345600;

export const CreateShareBody = zod.object({
  encryptedData: zod
    .string()
    .describe("Base64-encoded encrypted payload (ChaCha20-Poly1305)"),
  ttl: zod
    .number()
    .min(createShareBodyTtlMin)
    .max(createShareBodyTtlMax)
    .describe("Time-to-live in seconds (60 to 345600)"),
  passwordHash: zod
    .string()
    .nullish()
    .describe(
      "Optional PBKDF2-derived hash of the password (for verification)",
    ),
  passwordSalt: zod
    .string()
    .nullish()
    .describe("Base64-encoded salt used for password hashing"),
  webhookUrl: zod.string().nullish().describe("Optional HTTPS webhook URL"),
  webhookMessage: zod
    .string()
    .nullish()
    .describe("Optional custom webhook message"),
  fileMetadata: zod
    .array(
      zod.object({
        name: zod.string(),
        size: zod.number(),
        type: zod.string(),
        originalIndex: zod.number(),
      }),
    )
    .nullish(),
  shareType: zod.enum(["text", "files", "mixed"]),
  totalSize: zod.number().describe("Total size of the payload in bytes"),
  captchaToken: zod
    .string()
    .nullish()
    .describe("hCaptcha token (optional in dev mode)"),
});

/**
 * Fetch encrypted data for a share. Marks it as accessed immediately.
 * @summary Retrieve a share (one-time access)
 */
export const getSharePathShareIdMin = 64;
export const getSharePathShareIdMax = 64;

export const GetShareParams = zod.object({
  shareId: zod.coerce
    .string()
    .min(getSharePathShareIdMin)
    .max(getSharePathShareIdMax),
});

export const GetShareResponse = zod.object({
  encryptedData: zod.string(),
  fileMetadata: zod
    .array(
      zod.object({
        name: zod.string(),
        size: zod.number(),
        type: zod.string(),
        originalIndex: zod.number(),
      }),
    )
    .nullish(),
  passwordRequired: zod.boolean(),
  passwordSalt: zod.string().nullish(),
  shareType: zod.enum(["text", "files", "mixed"]),
  totalSize: zod.number(),
  webhookUrl: zod.string().nullish(),
  webhookMessage: zod.string().nullish(),
});

/**
 * Called by receiver after all files downloaded / text copied
 * @summary Delete a share after download complete
 */
export const deleteSharePathShareIdMin = 64;
export const deleteSharePathShareIdMax = 64;

export const DeleteShareParams = zod.object({
  shareId: zod.coerce
    .string()
    .min(deleteSharePathShareIdMin)
    .max(deleteSharePathShareIdMax),
});

export const DeleteShareResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * Returns size and whether password is required — does NOT mark as accessed
 * @summary Peek at share metadata without consuming it
 */
export const peekSharePathShareIdMin = 64;
export const peekSharePathShareIdMax = 64;

export const PeekShareParams = zod.object({
  shareId: zod.coerce
    .string()
    .min(peekSharePathShareIdMin)
    .max(peekSharePathShareIdMax),
});

export const PeekShareResponse = zod.object({
  totalSize: zod.number(),
  passwordRequired: zod.boolean(),
  shareType: zod.enum(["text", "files", "mixed"]),
  fileCount: zod.number(),
  expiresAt: zod.string(),
});

/**
 * @summary Test a webhook URL
 */
export const TestWebhookBody = zod.object({
  webhookUrl: zod.string(),
});

export const TestWebhookResponse = zod.object({
  success: zod.boolean(),
  statusCode: zod.number().nullish(),
  message: zod.string(),
});