# lib/api-zod/src/generated/types/createShareRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Secure Ephemeral Data Exchange Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateShareRequestShareType } from "./createShareRequestShareType";
import type { FileMetadata } from "./fileMetadata";

export interface CreateShareRequest {
  /** Base64-encoded encrypted payload (ChaCha20-Poly1305) */
  encryptedData: string;
  /**
   * Time-to-live in seconds (60 to 345600)
   * @minimum 60
   * @maximum 345600
   */
  ttl: number;
  /** Optional PBKDF2-derived hash of the password (for verification) */
  passwordHash?: string | null;
  /** Base64-encoded salt used for password hashing */
  passwordSalt?: string | null;
  /** Optional HTTPS webhook URL */
  webhookUrl?: string | null;
  /** Optional custom webhook message */
  webhookMessage?: string | null;
  fileMetadata?: FileMetadata[] | null;
  shareType: CreateShareRequestShareType;
  /** Total size of the payload in bytes */
  totalSize: number;
  /** hCaptcha token (optional in dev mode) */
  captchaToken?: string | null;
}