# lib/api-zod/src/generated/types/deleteShareResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Secure Ephemeral Data Exchange Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DeleteShareResponse {
  success: boolean;
  message: string;
}