# lib/api-zod/src/generated/types/detailedHealthStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Secure Ephemeral Data Exchange Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DetailedHealthStatus {
  status: string;
  containers_created: number;
  containers_delivered: number;
  memory_used_mb: number;
  memory_free_mb: number;
  timestamp: string;
}