# lib/api-zod/src/generated/types/fileMetadata.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Secure Ephemeral Data Exchange Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface FileMetadata {
  name: string;
  size: number;
  type: string;
  originalIndex: number;
}