# lib/api-zod/src/generated/types/getShareResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Secure Ephemeral Data Exchange Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { FileMetadata } from "./fileMetadata";
import type { GetShareResponseShareType } from "./getShareResponseShareType";

export interface GetShareResponse {
  encryptedData: string;
  fileMetadata?: FileMetadata[] | null;
  passwordRequired: boolean;
  passwordSalt?: string | null;
  shareType: GetShareResponseShareType;
  totalSize: number;
  webhookUrl?: string | null;
  webhookMessage?: string | null;
}