# lib/api-zod/src/generated/types/testWebhookResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Secure Ephemeral Data Exchange Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TestWebhookResponse {
  success: boolean;
  statusCode?: number | null;
  message: string;
}